Reorganize project, add h2 dependency
