--- src/main/java/com/mashanlote/model/WeatherPOJO.java
@@ -1,3 +1,0 @@
-package com.mashanlote.model;
-
-public record WeatherPOJO(CurrentWeather current, Location location) { }
