+++ src/main/java/com/mashanlote/model/weather/CreateRegionRequest.java
@@ -1,3 +1,3 @@
-package com.mashanlote.model;
+package com.mashanlote.model.weather;
 
 public record CreateRegionRequest(String regionName, int temperature) { }
