+++ src/main/java/com/mashanlote/model/weather/ErrorDetails.java
@@ -1,3 +1,3 @@
-package com.mashanlote.model;
+package com.mashanlote.model.weather;
 
 public record ErrorDetails(String code, String message) { }
