+++ src/main/java/com/mashanlote/model/weather/WeatherUpdate.java
@@ -1,3 +1,3 @@
-package com.mashanlote.model;
+package com.mashanlote.model.weather;
 
 public record WeatherUpdate(int temperature) { }
