+++ src/main/java/com/mashanlote/model/weatherapi/ApiError.java
@@ -1,4 +1,4 @@
-package com.mashanlote.model;
+package com.mashanlote.model.weatherapi;
 
 public record ApiError(ApiErrorInternal error) { }
