+++ src/main/java/com/mashanlote/model/weatherapi/ApiErrorInternal.java
@@ -1,3 +1,3 @@
-package com.mashanlote.model;
+package com.mashanlote.model.weatherapi;
 
 public record ApiErrorInternal(Integer code, String message) { }
