+++ src/main/java/com/mashanlote/model/weatherapi/Condition.java
@@ -1,3 +1,3 @@
-package com.mashanlote.model;
+package com.mashanlote.model.weatherapi;
 
 public record Condition(Integer code, String icon, String text) { }
