+++ src/main/java/com/mashanlote/model/weatherapi/WeatherDTO.java
@@ -1,0 +1,3 @@
+package com.mashanlote.model.weatherapi;
+
+public record WeatherDTO(CurrentWeather current, Location location) { }
