+++ src/main/resources/application.properties
@@ -3,3 +3,5 @@ api-key=${WEATHER_API_KEY}
 resilience4j.ratelimiter.instances.api.limit-refresh-period=30d
 resilience4j.ratelimiter.instances.api.timeout-duration=3s
 resilience4j.ratelimiter.instances.api.limit-for-period=800000
+spring.h2.console.enabled=true
+spring.datasource.url=jdbc:h2:mem:test
